+++ budget-app/src/components/ExpenseList.jsx
@@ -1,16 +1,20 @@
 import React, { useState } from 'react';
-import {
-  Input,
-  Button,
-  Flex,
-  NumberInput,
-  NativeSelect,
-} from '@chakra-ui/react';
+import { v4 as uuidv4 } from 'uuid';
 
 const ExpenseList = ({ items, setItems, places }) => {
-  const [formData, setFormData] = useState({ place: '', price: '', date: '' });
-  const [editIndex, setEditIndex] = useState(null);
-  const [editData, setEditData] = useState({ place: '', price: '', date: '' });
+  const [formData, setFormData] = useState({
+    id: '',
+    place: '',
+    price: '',
+    date: '',
+  });
+  const [editId, setEditId] = useState(null);
+  const [editData, setEditData] = useState({
+    id: '',
+    place: '',
+    price: '',
+    date: '',
+  });
 
   const handleChange = (e) => {
     setFormData({ ...formData, [e.target.name]: e.target.value });
@@ -19,41 +23,42 @@ const ExpenseList = ({ items, setItems, places }) => {
   const handleAdd = (e) => {
     e.preventDefault();
     if (!formData.place || !formData.price || !formData.date) return;
-    setItems([...items, formData]);
-    setFormData({ place: '', price: '', date: '' });
+    const newItem = { id: uuidv4(), ...formData }; // ✅ IDを付ける
+    setItems([...items, newItem]);
+    setFormData({ place: '', price: '', date: '' }); // idはリセット不要
   };
 
-  const handleEditClick = (index) => {
-    setEditIndex(index);
-    setEditData(items[index]);
+  const handleEditClick = (id) => {
+    const target = items.find((item) => item.id === id);
+    setEditId(id);
+    setEditData(target);
   };
 
   const handleEditChange = (e) => {
     setEditData({ ...editData, [e.target.name]: e.target.value });
   };
 
-  const handleSave = (index) => {
-    const updated = [...items];
-    updated[index] = editData;
+  const handleSave = () => {
+    const updated = items.map((item) =>
+      item.id === editId ? { ...editData, id: editId } : item
+    );
     setItems(updated);
-    setEditIndex(null);
+    setEditId(null);
   };
 
   const handleCancel = () => {
-    setEditIndex(null);
+    setEditId(null);
   };
 
-  const handleDelete = (index) => {
-    setItems(items.filter((_, i) => i !== index));
-    if (editIndex === index) setEditIndex(null);
+  const handleDelete = (id) => {
+    setItems(items.filter((item) => item.id !== id));
+    if (editId === id) setEditId(null);
   };
 
   return (
-    <div style={{ marginTop: '20px' }}>
+    <div>
       <form onSubmit={handleAdd}>
         <input
-          size="xs"
-          width="150px"
           type="date"
           name="date"
           value={formData.date}
@@ -70,89 +75,57 @@ const ExpenseList = ({ items, setItems, places }) => {
           ))}
         </select>
         <input
+          type="number"
           name="price"
           value={formData.price}
           onChange={handleChange}
           placeholder="支出金額"
         />
-        円
-        <button size="xs" colorPalette="blue" type="submit">
-          ＋
-        </button>
+        円<button type="submit">＋</button>
       </form>
 
       <h3>入力一覧</h3>
       <ul>
         {[...items]
           .sort((a, b) => new Date(a.date) - new Date(b.date))
-          .map((item, index) => (
-            <li key={index}>
-              {editIndex === index ? (
-                <select>
+          .map((item) => (
+            <li key={item.id}>
+              {editId === item.id ? (
+                <>
                   <input
-                    size="xs"
-                    width="150px"
                     type="date"
                     name="date"
                     value={editData.date}
                     onChange={handleEditChange}
                   />
-
-                    <select
-                      name="place"
-                      value={editData.place}
-                      onChange={handleEditChange}
-                    >
-                      <option value="" disabled>
-                        場所を選択
+                  <select
+                    name="place"
+                    value={editData.place}
+                    onChange={handleEditChange}
+                  >
+                    <option value="" disabled>
+                      場所を選択
+                    </option>
+                    {places.map((place, idx) => (
+                      <option key={idx} value={place}>
+                        {place}
                       </option>
-                      {places.map((place, idx) => (
-                        <option key={idx} value={place}>
-                          {place}
-                        </option>
-                      ))}
-                    </select>
-
+                    ))}
+                  </select>
                   <input
                     name="price"
                     value={editData.price}
                     onChange={handleEditChange}
                     placeholder="支出金額"
                   />
-                  円
-                  <button
-                    size="xs"
-                    colorPalette="blue"
-                    onClick={() => handleSave(index)}
-                  >
-                    保存
-                  </button>
-                  <button
-                    size="xs"
-                    colorPalette="gray"
-                    variant="subtle"
-                    onClick={handleCancel}
-                  >
-                    キャンセル
-                  </button>
+                  円<button onClick={handleSave}>保存</button>
+                  <button onClick={handleCancel}>キャンセル</button>
                 </>
               ) : (
                 <>
                   {item.date}｜{item.place}：{item.price} 円
-                  <button
-                    size="xs"
-                    colorPalette="green"
-                    onClick={() => handleEditClick(index)}
-                  >
-                    ↺
-                  </button>
-                  <button
-                    size="xs"
-                    colorPalette="red"
-                    onClick={() => handleDelete(index)}
-                  >
-                    ー
-                  </button>
+                  <button onClick={() => handleEditClick(item.id)}>↺</button>
+                  <button onClick={() => handleDelete(item.id)}>ー</button>
                 </>
               )}
             </li>
